separate command and special security mantles
Gives most command members a mantle suitable for their role usually with most of the stats of a mantle in the department but with a command contraband component.

It gives corpsman's mantle security hud and flash protection and gives the detectives mantle a civilian appearance and as noir overlay.

In addition it returns flash protection to Chief of Medical, Corpsman, Research Director and Captain, and provides the Director's Mantle with the binary hivemind and what should be diaghud, but i think its just a slightly weaker medical mantle hud.

I think it also makes NTR and AA spawn in with command mantles hopefully.

I was gonna make more mantles specifically for bartender, paramedics, roboticist and chemist but i cant seem to figure out how to give mantles solution analysers...the roboticist isnt mapped, the bartenders hud I belive isnt in the game yet and Im worried that whenever I get complacent and touch the meta.json file for any reason at all I fat finger another fatal error into the changes.